+++ torch_geometric/sampler/base.py
@@ -179,7 +179,7 @@ class HeteroSamplerOutput(CastMixin):
 
 
 @dataclass(frozen=True)
-class NumNeighbors(CastMixin):
+class NumNeighbors:
     r"""The number of neighbors to sample in a homogeneous or heterogeneous
     graph. In heterogeneous graphs, may also take in a dictionary denoting
     the amount of neighbors to sample for individual edge types.
@@ -277,6 +277,10 @@ class NumNeighbors(CastMixin):
 
         self.__dict__['_num_hops'] = num_hops
         return num_hops
+
+    def __len__(self) -> int:
+        r"""Returns the number of hops."""
+        return self.num_hops
 
 
 class NegativeSamplingMode(Enum):